Fixed test file dependency bug.
The TEST_TARGET dependency rule had a dependency on the static library
(LIB_TARGET) but the link command in the TEST_TARGET rule uses "-lrmc" to link
the library which uses the dynamic library preferentially if it exists.  The
dependency should be for the LIB_SO_TARGET as this is what will actually be
used by the linker.

The result of this bug is that there is a race condition between the make
thread that is linking the librmc.so file and the thread that is linking the
rmc_test executable.  If the librmc.so finishes first, there is no problem and
everything works fine.  But if the librmc_test executable tries to use the
librmc.so file before it is finished being built, an error occurs and the link
fails.

By fixing the dependency rule, we guarantee that the librmc.so library will
always be finished before the rmc_test executable tries to use it to link
with.